NumberDropdown.js を tsx に書き換えました。
また、汎用性があると考えたため components/ に移動しました。
